+++ android/src/main/java/com/jhotadhari/reactnative/mapsforge/vtm/react/modules/MapContainerModule.java
@@ -274,13 +274,20 @@ public class MapContainerModule extends ReactContextBaseJavaModule {
 	}
 
 	@ReactMethod
-	public void setMapEventRate(int nativeNodeHandle, int mapEventRate, Promise promise ) {
+	public void setRate(int nativeNodeHandle, String key, int newRate, Promise promise ) {
 		try {
 			MapFragment mapFragment = (MapFragment) Utils.getMapFragment( this.getReactApplicationContext(), nativeNodeHandle );
 			if ( null == mapFragment ) {
 				promise.reject( "Error", "Unable to find mapFragment" ); return;
 			}
-			mapFragment.updateRateLimiterRate( mapEventRate );
+			switch( key ) {
+				case "mapEventRate":
+					mapFragment.updateRateLimiterRate( newRate );
+					break;
+				case "hgtReadFileRate":
+					mapFragment.updateHgtReadFileRate( newRate );
+					break;
+			}
 			promise.resolve( true );
 		} catch( Exception e ) {
 			e.printStackTrace();